Sunrise Plugin Compatibility
Add notes in sunrise.php about which plugins this has been tested with.
This includes their name and a link to their page in the WordPress
repository.
**Please note, this solution works with WP Multi Network, however that
plugin has its own bug where moving a site from one network to another
will cause all the sites in the destination network, except the new
network's root site, to be orphaned and not accessible from WordPress.